refactor: extract transaction receipt view from split-app
